refactor: some cleanup
