--- packages/rich-text/src/plugins/Table/components/TableActions/index.ts
@@ -1,1 +1,0 @@
-export * from './TableActions';
