--- packages/rich-text/src/plugins/Table/components/TableActions/styles.ts
@@ -1,10 +1,0 @@
-import { css } from 'emotion';
-import tokens from '@contentful/forma-36-tokens';
-
-export const styles = {
-  topRight: css({
-    position: 'absolute',
-    top: tokens.spacingXs,
-    right: tokens.spacingXs,
-  }),
-};
